+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-### V1.6.0
+### V1.0.7
+- Added auto enable familiar equip emotes option (enabled by default)
+- Fixed windows to hide when escape menu is open
+- Fixed some message parsing overlap with Legacy module
+### V1.0.6
 - Updated for V Rising version 1.1.8 & BloodCraft 1.9.7
 - Removed Bloodstone dependency